refactored sql code, refactored vulnerable queries.

+++ src/main/java/com/excercise/UserInterface.java
@@ -9,35 +9,43 @@ import java.util.Scanner;
 public class UserInterface {
 
     public static void StartUI() throws Exception {
-         String commands="\n1 - Add a game to collection \n2 - Registrate a play of a game\n3 - choose a random game to play\n4- List all games in collection\n5 - Statstics of a selected game\n6- delete a single game from library\n7-delete all games from library\n0-exits app";
+        String commands = "\n1 - Add a game to collection \n2 - Registrate a play of a game\n3 - choose a random game to play\n4- List all games in collection\n5 - Statstics of a selected game\n6- delete a single game from library\n7-delete all games from library\n0-exits app";
         System.out.println(commands);
-         Scanner scanner= new Scanner(System.in);
-         String cmd=scanner.nextLine();
-         switch (cmd){
-             case "1":AddGame();
-             break;
-             case "2":AddPlay();
-             break;
-             case "3":ChooseRandomGame();
-             break;
-             case "4":ListAllGames();
-             break;
-             case "5":StatsOfSelectedGame();
-             break;
-             case "6":RemoveGame();
-             break;
-             case "7":DeleteAllGames();
-             break;
-             case "0":break;
-         }
+        Scanner scanner = new Scanner(System.in);
+        String cmd = scanner.nextLine();
+        switch (cmd) {
+            case "1":
+                AddGame();
+                break;
+            case "2":
+                AddPlay();
+                break;
+            case "3":
+                ChooseRandomGame();
+                break;
+            case "4":
+                ListAllGames();
+                break;
+            case "5":
+                StatsOfSelectedGame();
+                break;
+            case "6":
+                RemoveGame();
+                break;
+            case "7":
+                DeleteAllGames();
+                break;
+            case "0":
+                break;
+        }
     }
+
     @SneakyThrows
     public static void AddGame() {
-        Scanner scanner= new Scanner(System.in);
+        Scanner scanner = new Scanner(System.in);
         System.out.println("Enter title of game:");// check for uniqueness!!
-        String cmd =scanner.nextLine().toLowerCase().trim();
-//        IsGameInLibrary(cmd);
-        Boardgame game=new Boardgame();
+        String cmd = scanner.nextLine().toLowerCase().trim();
+        Boardgame game = new Boardgame();
         game.setName(cmd);
         System.out.println("Enter type of game:");
         game.setGameType(scanner.nextLine().toLowerCase().trim());
@@ -52,32 +60,36 @@ public class UserInterface {
 
     @SneakyThrows
     public static void AddPlay() throws Exception {
-        Scanner scanner= new Scanner(System.in);
+        Scanner scanner = new Scanner(System.in);
         System.out.println("Enter the title of the game");// TODO: implement "or a part of it"
-        String cmd=scanner.nextLine();
-        IsGameInLibrary(cmd);
-        Boardgame boardgame=new Boardgame(cmd);
+        String cmd = scanner.nextLine();
+        if (!IsGameInLibrary(cmd)) {
+            System.out.println("Game not found.");
+            return;
+        }
+        Boardgame boardgame = new Boardgame(cmd);
         BoardGameRepository.update(boardgame);
         System.out.println("Update succesful. Returning to main menu.");
         StartUI();
     }
+
     @SneakyThrows
-    public static void ChooseRandomGame (){
+    public static void ChooseRandomGame() {
         List<Boardgame> boardgames = BoardGameRepository.findAll();
-        if(boardgames.isEmpty()){
+        if (boardgames.isEmpty()) {
             System.out.println("The list is empty. Probably should buy some games");
             StartUI();
         }
-        Random rnd=new Random();
-        Boardgame randomBoard=boardgames.get(rnd.nextInt(boardgames.size()));
-        System.out.println("Your random boardgame is:" + randomBoard.getName()+".\nReturning to main menu.");
+        Random rnd = new Random();
+        Boardgame randomBoard = boardgames.get(rnd.nextInt(boardgames.size()));
+        System.out.println("Your random boardgame is:" + randomBoard.getName() + ".\nReturning to main menu.");
         StartUI();
 
     }
 
-    public static void  ListAllGames() throws Exception {
+    public static void ListAllGames() throws Exception {
         List<Boardgame> boardgames = BoardGameRepository.findAll();
-        if(boardgames.isEmpty()){
+        if (boardgames.isEmpty()) {
             System.out.println("The list is empty. Probably should buy some games. Returning to main menu");
             StartUI();
         }
@@ -85,34 +97,36 @@ public class UserInterface {
             System.out.println(x.getName());
         }
         System.out.println("To return to main menu, enter MAIN below, to exit - press any button");
-        Scanner scanner= new Scanner(System.in);
-        if (scanner.nextLine().toLowerCase().trim().equals("main")){
+        Scanner scanner = new Scanner(System.in);
+        if (scanner.nextLine().toLowerCase().trim().equals("main")) {
             StartUI();
         }
     }
+
     @SneakyThrows
     public static void StatsOfSelectedGame() throws Exception {
         System.out.println("Enter a game title:");
         Scanner scanner = new Scanner(System.in);
 
-            String title = scanner.nextLine().trim();
-            Boardgame game = BoardGameRepository.findOne(title);
-
-            if (game == null) {
-                System.out.println("Game not found. Returning to main menu.");
-                return;
-            }
-
-            System.out.println("Statistics of selected game: " + game);
+        String title = scanner.nextLine().trim();
+        Boardgame game = BoardGameRepository.findOne(title);
+        if (!IsGameInLibrary(title)) {
+            System.out.println("Game not found.");
+            return;
         }
-    }
 
+
+        System.out.println("Statistics of selected game: " + game);
+    }
     @SneakyThrows
     public static void  RemoveGame(){
         System.out.println("Enter a game title:");
         Scanner scanner= new Scanner(System.in);
         String title=scanner.nextLine();
-        IsGameInLibrary(title);
+        if (!IsGameInLibrary(title)) {
+            System.out.println("Game not found.");
+            return;
+        }
         BoardGameRepository.deleteOne(title);
     }
 
@@ -130,11 +144,6 @@ public class UserInterface {
     @SneakyThrows
     public static boolean IsGameInLibrary(String name){
         List<Boardgame> boardgames = BoardGameRepository.findAll();
-//        if(boardgames.isEmpty()){
-//            System.out.println("The list is empty. Probably should buy some games");
-//            System.out.println("Returning to initial menu");
-//            StartUI();
-//        }
         return boardgames.stream().
                 map(Boardgame::getName)
                 .anyMatch(gameName->gameName.contains(name));
